Heartbeat with xDrip4iOS
- add option to make bluetooth connection to CGM transmitter, which will then work as heartbeat.
Enabled by default. Can be disabled in settings.
A call is made to fetNewDataIfNeeded each time there's a call to 'peripheral(_:didUpdateValueFor:error:)' or centralManager(_:didDisconnectPeripheral:error:)
The call on didDisconnect is useful for Dexcom users.

- removed DispatchTimer. Timers don't work when the app is in the background. App will soon be suspended after going to background and timers are triggered. The Bluetooth trigger mechanism is what will wake up the app and trigger a fetch of the last reading

- possibility to send trace info via e-mail. This is only for xdrip-client functionality. Can be removed later.

- in xDripClientManager.fetchNewDataIfNeeded :  replaced all
   self.delegateQueue.async { completion( .. data ..) }
     By
   self.delegate.notify { (delegate) in delegate?.cgmManager(self, didUpdateWith: .noData) }
Because in case fetchNewDataIfNeeded due to wake up signal by BluetoothTransmitter, then there's no completion handler available. Seems to work well.
And I'm not doing this async, this was causing Loop failures

- add observer on UIApplication.didFinishLaunchingNotification and UIApplication.willEnterForegroundNotification : a call to fetNewDataIfNeeded is made when triggered